backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
PyPDF2==3.0.1
python-dotenv==1.0.0